requirements.txt
azure.functions
azure-functions-worker
requests